requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
pyyaml==6.0.1
pytest-playwright

